Changed: Refactor model catalog API for cleaner public interface
Consolidate builder input types and improve internal encapsulation.

Changes:
- Merged ModelInfo and ModelConfig into single ModelInfo struct with Option<f32> for sampling params
- Moved ModelCatalogBuilder to catalog level (from internal/)
- Moved Modality to public/ folder for external access
- Extracted ModelCatalog to dedicated catalog.rs file
- Model and CatalogEntry now store Fixed4 internally, expose Option<f32> via accessors
- Marked test-only Fixed4 methods with #[cfg(test)] instead of blanket #[allow(dead_code)]
- Added public API documentation to module header

Benefits:
- Users no longer need to understand Fixed4 encoding - just pass Option<f32>
- Cleaner module structure separates public types from internal implementation
- Size-optimized internal storage (Fixed4 = 2 bytes) with ergonomic public API
- All catalog types exported from single module with wildcard re-export
