+++ src/llm-coding-tools-core/src/models/catalog/public/mod.rs
@@ -1,14 +1,16 @@
-//! Public types and APIs for the model catalog.
+//! Public types for the model catalog.
 //!
-//! This module contains only the public-facing types and the builder
-//! needed to construct a [`ModelCatalog`].
+//! This module contains public-facing data types used when building
+//! and querying a [`ModelCatalog`].
 
-pub use builder_types::{LookupTableKind, ModelCatalogBuildError, ProviderInfo};
+pub use builder_types::{LookupTableKind, ModelCatalogBuildError, ModelInfo, ProviderInfo};
 pub use entry::{CatalogEntry, Model, Provider};
+pub use modality::Modality;
 pub use model_idx::ModelIdx;
 pub use provider_idx::ProviderIdx;
 
 pub(crate) mod builder_types;
 mod entry;
+mod modality;
 mod model_idx;
 mod provider_idx;
