+++ src/llm-coding-tools-core/src/models/catalog/public/modality.rs
